fix: mock useWeb3React
