+++ src/components/PositionListItem/PositionListItem.test.tsx
@@ -20,16 +20,6 @@ jest.mock('hooks/Tokens')
 // eslint-disable-next-line react/display-name
 jest.mock('components/DoubleLogo', () => () => <div />)
 
-jest.mock('@web3-react/core', () => {
-  const web3React = jest.requireActual('@web3-react/core')
-  return {
-    ...web3React,
-    useWeb3React: () => ({
-      chainId: 1,
-    }),
-  }
-})
-
 const susToken0Address = '0x39AA39c021dfbaE8faC545936693aC917d5E7563'
 
 beforeEach(() => {
