+++ src/components/SearchModal/CurrencyList/index.test.tsx
@@ -23,17 +23,6 @@ jest.mock(
       `CurrencyLogo currency=${currency.symbol}`
 )
 
-jest.mock('@web3-react/core', () => {
-  const web3React = jest.requireActual('@web3-react/core')
-  return {
-    ...web3React,
-    useWeb3React: () => ({
-      account: '123',
-      isActive: true,
-    }),
-  }
-})
-
 jest.mock('../../../state/connection/hooks', () => {
   return {
     useCurrencyBalance: (currency: Currency) => {
