+++ src/pages/Pool/CTACards.test.tsx
@@ -3,16 +3,6 @@ import { render, screen } from 'test-utils/render'
 
 import CTACards from './CTACards'
 
-jest.mock('@web3-react/core', () => {
-  const web3React = jest.requireActual('@web3-react/core')
-  return {
-    ...web3React,
-    useWeb3React: () => ({
-      chainId: 42,
-    }),
-  }
-})
-
 describe('CTAcard links', () => {
   it('renders mainnet link when chain is not supported', () => {
     jest.spyOn(useV3Positions, 'useV3Positions').mockImplementation(() => {
